Tidied, at least a little
