+++ config.py
@@ -1,0 +1,2 @@
+TOKEN      = "TOKEN"
+USER_AGENT = "USER_AGENT"
